Update Address Commands

+++ src/Application/Addresses/Commands/CreateAddress/CreateAddressCommand.cs
@@ -19,8 +19,8 @@ namespace DeliveryWebApp.Application.Addresses.Commands.CreateAddress
         public string PostalCode { get; set; }
         public string StateProvince { get; set; }
         public string Country { get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Latitude { get; set; }
         public int CustomerId { get; set; }
         public Customer Customer { get; set; }
     }
